ajouter des enums à l'entité Customer
