+++ True.Delta/True.Delta-ejb/src/main/java/Enumerations/Active_account_Type.java
@@ -1,0 +1,5 @@
+package Enumerations;
+
+public enum Active_account_Type {
+ Active,Not_Active
+}
