+++ True.Delta/True.Delta-ejb/src/main/java/Enumerations/Diversified_Type.java
@@ -1,0 +1,5 @@
+package Enumerations;
+
+public enum Diversified_Type {
+	Diversified, Not_Diversified
+}
