+++ True.Delta/True.Delta-ejb/src/main/java/Enumerations/Professional_Status_Type.java
@@ -1,0 +1,5 @@
+package Enumerations;
+
+public enum Professional_Status_Type {
+	Doctor,Professor,Engineer,other
+}
