+++ True.Delta/True.Delta-ejb/src/main/java/Enumerations/Residency_Status_Type.java
@@ -1,0 +1,6 @@
+package Enumerations;
+
+
+public enum Residency_Status_Type {
+		Owner,Tenant
+}
