+++ True.Delta/True.Delta-ejb/src/main/java/Enumerations/Type_of_contract_type.java
@@ -1,0 +1,6 @@
+package Enumerations;
+
+public enum Type_of_contract_type {
+
+The_Initiation_into_the_World_of_Work_Scheme,Fixed_term_contract,permanent_work_contract
+}
